reworked structure of application subpackage
