--- fileformats/extras/application/archive/__init__.py
@@ -1,1 +1,0 @@
-from .converters import *
