--- fileformats/extras/application/serialization/__init__.py
@@ -1,2 +1,0 @@
-from . import converters
-from . import yaml
